Added props to SectionWrapper
- added all possible background-colors from the design file (SECONDARY_CREAM, SECONDARY_CORNFLOWER, PRIMARY_CORAL, VIOLET_8) as props
- added a default and hero padding that is responsive. Add new classes in the scss file if a section doesn't follow the default padding (ASK ANCA IF YOU DON'T KNOW HOW)
